Move make_map_counts to counts.py
