+++ gammapy/cube/counts.py
@@ -2,9 +2,11 @@
 """Functions to perform basic functions for map and cube analysis.
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
+from ..maps import WcsNDMap
 
 __all__ = [
     'fill_map_counts',
+    'make_map_counts',
 ]
 
 
@@ -42,3 +44,35 @@ def fill_map_counts(count_map, event_list):
                 raise ValueError("Cannot find MapGeom axis {!r} in EventList".format(axis.name))
 
     count_map.fill_by_coord(coord_dict)
+
+
+def make_map_counts(events, ref_geom, pointing, offset_max):
+    """Build a WcsNDMap (space - energy) with events from an EventList.
+
+    The energy of the events is used for the non-spatial axis.
+
+    Parameters
+    ----------
+    events : `~gammapy.data.EventList`
+        Event list
+    ref_geom : `~gammapy.maps.WcsGeom`
+        Reference WcsGeom object used to define geometry (space - energy)
+    pointing : `~astropy.coordinates.SkyCoord`
+        Pointing direction
+    offset_max : `~astropy.coordinates.Angle`
+        Maximum field of view offset.
+
+    Returns
+    -------
+    cntmap : `~gammapy.maps.WcsNDMap`
+        Count cube (3D) in true energy bins
+    """
+    counts_map = WcsNDMap(ref_geom)
+    fill_map_counts(counts_map, events)
+
+    # Compute and apply FOV offset mask
+    offset = ref_geom.separation(pointing)
+    offset_mask = offset >= offset_max
+    counts_map.data[:, offset_mask] = 0
+
+    return counts_map
